edit about us

+++ database/migrations/2017_01_02_202412_create_ratings_table.php
@@ -17,7 +17,7 @@ class CreateRatingsTable extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned()->foreign('user_id')->reference('id')->on('users')->onDelete('cascade');
             $table->integer('movie_id')->unsigned()->foreign('movie_id')->reference('id')->on('movies')->onDelete('cascade');
-            $table->float('rating');
+            $table->decimal('rating');
             $table->rememberToken();
             $table->timestamps();
         });
